XP12.06 changed `draw_call_type` in window callbacks, use OGL matrices. (#23)
* XP12.06 changed `draw_call_type` in window callbacks, use OGL matrices.

In mt_cairo_render, using the viewport datarefs in window callbacks doesn't
scale with the interface scaling setting. Before we used the matrices that
XP gives us instead, because the draw_call_type dataref was (wrongly) set
to 1 in window callbacks. We fix it by _always_ using the matrices instead
of the viewport XP provides us.